+++ CHANGELOG.md
@@ -14,6 +14,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - check wazuh version
 - assets tab, show trash problems. Not when search::show but then sort problems ????
 - upsert could be the wrong way because of loosing items history
+- agent sync lower timeout
+
+## [0.0.5] 2025-04-05
+### Added
+- Default ticket ITIL Category
 
 ## [0.0.3] 2025-04-04